requirements.txt
coverage==3.6
Django==2.2.2
django-annoying==0.10.4
django-environ==0.4.5
django-user-agents==0.3.2
django-widget-tweaks==1.4.3
psycopg2==2.8.3
python-memcached==1.59
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.3.0
ua-parser==0.8.0
urllib3==1.25.3
user-agents==2.0
